feat(nested-menu): show menu option as nested menu if case
Update the interface for MenuOption to include an optional field for a
nested menu.
Update the pug template and the style file to show an arrow in case a
menu option includes a sub menu.
Add a nested menu option in the demo for testing.
Update the documentation to include nestedMenu property.